Split examples to separate files.
