+++ mapper/cumulocity/c8y_translator_lib/examples/single_value_translation.rs
@@ -1,0 +1,24 @@
+use c8y_translator_lib::json::from_thin_edge_json;
+
+fn single_value_translation() {
+    let single_value_thin_edge_json = r#"  {
+    "temperature": 23,
+    "pressure": 220
+   }"#;
+
+    println!("Thin_Edge_Json: \n{:#}", single_value_thin_edge_json);
+
+    let output = from_thin_edge_json(single_value_thin_edge_json.as_bytes());
+    match output {
+        Ok(vec) => {
+            println!("{}", String::from_utf8(vec).unwrap());
+        }
+        Err(e) => {
+            eprintln!("Error: {}", e);
+        }
+    }
+}
+
+pub fn main() {
+    single_value_translation();
+}
